fixed dir some components
