+++ src/Profile/PersonInfo.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
-import iconCheck from './check.svg';
+import iconCheck from '../Recommendations/check.svg';
 import Button from '../UI/Button';
 
 const Profile = styled.div`
