+++ src/Trends/index.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import Header from './Header';
+import Trends from './Trends';
+
+export default () => (
+  <React.Fragment>
+    <Header>United Kingdom Trends</Header>
+    <Trends />
+  </React.Fragment>
+);
